Add account area scaffolding: single /account page with nav + section routing
Fase 0: nuevo AccountPage (gating + barra contexto + sidebar/tabs con sección
por query param), tokens de tema fondoApp/exito, migración de las 5 secciones
actuales sin rediseñar, y redirecciones de las rutas viejas a /account.
