+++ src/pages/account.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import AccountPage from '../views/AccountPage';
+import { PublicZoneLayout } from '../components/Layouts';
+
+const Account: React.FC = (): JSX.Element => (
+  <PublicZoneLayout>
+    <AccountPage />
+  </PublicZoneLayout>
+);
+
+export default Account;
