+++ src/styled.d.ts
@@ -15,5 +15,7 @@ declare module 'styled-components' {
     brown: string;
     white: string;
     lightBorder: string;
+    fondoApp: string;
+    exito: string;
   }
 }
