+++ src/views/AccountPage/sections/AddBookSection.tsx
@@ -1,0 +1,346 @@
+/* eslint-disable no-underscore-dangle */
+import React, { useState, useContext, useEffect } from 'react';
+import { useRouter } from 'next/router';
+import styledComponents from 'styled-components';
+import { useTranslation } from 'react-i18next';
+import {
+  Card,
+  CardContent,
+  Button,
+  Collapse,
+  FormHelperText,
+  Grid,
+  IconButton,
+  TextField,
+} from '@mui/material';
+import CloseIcon from '@mui/icons-material/Close';
+import Alert from '@mui/material/Alert';
+import { AdapterDateFns } from '@mui/x-date-pickers/AdapterDateFns';
+import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider';
+import { DatePicker } from '@mui/x-date-pickers/DatePicker';
+import ReactGA from 'react-ga4';
+import {
+  useForm,
+  useUploadImages,
+  useRequiredFieldsValidation,
+  useFetchBook,
+  useFetch,
+} from '../../../utils/customHooks';
+import UserContext from '../../../store/context/userContext/UserContext';
+import { registerBook as URL } from '../../../config/routes';
+import { FormatsCheckBoxSelector, GenresSelector, Loading } from '../../../components';
+import { BookForm, BookFormErrors } from '../../../interfaces/books';
+import SectionHeader from '../SectionHeader';
+
+const AddBookSection: React.FC = (): JSX.Element => {
+  const { t } = useTranslation();
+  const { query: { book } } = useRouter();
+  const { user } = useContext(UserContext);
+  const [state, fetchBook] = useFetchBook();
+  const [registerBookResponse, registerBookRequest, loading] = useFetch();
+
+  const initForm: BookForm = {
+    title: '',
+    cover: '',
+    formats: [],
+    datePublished: '',
+    author: user._id,
+    synopsis: '',
+    genre: '',
+    pages: '',
+  };
+  const [bookForm, setBookForm, loadForm] = useForm(initForm);
+
+  const initErrors: BookFormErrors = {
+    title: '',
+    cover: '',
+    formats: '',
+    datePublished: '',
+    author: '',
+    synopsis: '',
+    editorial: '',
+    genre: '',
+    pages: '',
+  };
+  const [errors, validateRequiredFields] = useRequiredFieldsValidation(initErrors);
+  const [open, setOpen] = useState(false);
+
+  const [coverURL, uploadCover, loadingCover] = useUploadImages(initForm.cover);
+
+  useEffect(() => {
+    setBookForm('cover', coverURL);
+  }, [coverURL]);
+
+  useEffect(() => {
+    if (book) fetchBook(book);
+  }, [book]);
+
+  useEffect(() => {
+    const {
+      title,
+      synopsis,
+      cover,
+      editorial,
+      formats,
+      genre,
+      pages,
+      datePublished,
+    } = state;
+    loadForm({
+      ...initForm,
+      title,
+      synopsis,
+      cover,
+      editorial,
+      formats,
+      genre,
+      pages,
+      datePublished,
+    });
+  }, [state]);
+
+  useEffect(() => {
+    if (registerBookResponse !== undefined) setOpen(true);
+  }, [registerBookResponse.message]);
+
+  const registerBook = async (): Promise<void> => {
+    if (!book) {
+      registerBookRequest(URL, 'post', bookForm);
+      ReactGA.event({
+        category: 'Nuevo Libro',
+        action: `Libro ${bookForm.title}`,
+      });
+    } else {
+      registerBookRequest(`${URL}/${book}`, 'put', bookForm);
+    }
+  };
+
+  const submit = () => {
+    const requiredFields = ['title', 'synopsis', 'genre', 'cover', 'pages', 'formats', 'datePublished'];
+    validateRequiredFields(bookForm, requiredFields, registerBook);
+  };
+
+  return (
+    <>
+      <SectionHeader title={t('titles.addBook')} />
+      <Card>
+        <StyledFormContainer>
+          <StyledFirstColumnContainer>
+            <StyledInputFile
+              disabled={loadingCover}
+              accept="image/*"
+              id="raised-button-file"
+              multiple
+              type="file"
+              onChange={(e) => uploadCover(e, 'covers')}
+            />
+            <StyledLabel htmlFor="raised-button-file">
+              {loadingCover ? t('buttons.loading') : t('buttons.uploadCover')}
+            </StyledLabel>
+            <FormHelperText error>{errors.cover}</FormHelperText>
+            {
+              loadingCover
+                ? <Loading />
+                : bookForm.cover
+                && <StyledImageContainer src={bookForm.cover} alt="Cover" />
+            }
+          </StyledFirstColumnContainer>
+          <StyledSecondColumnContainer>
+            {['title', 'author', 'editorial'].map((text) => (
+              <TextField
+                label={t(`booksForm.${text}`)}
+                name={text}
+                fullWidth
+                onChange={({ target: { name, value } }) => setBookForm(name, value)}
+                variant={text === 'author' ? 'filled' : 'outlined'}
+                size="small"
+                required={text !== 'editorial'}
+                placeholder={text === 'editorial' ? 'Dejalo en blanco si eres autor independiente' : ''}
+                error={errors[text].length > 0}
+                helperText={errors[text]}
+                InputProps={{
+                  readOnly: (text === 'author'),
+                }}
+                value={text !== 'author' ? bookForm[text] : `${user.name} ${user.lastName || ''}`}
+              />
+            ))}
+            <TextField
+              id="standard-full-width"
+              label="Synopsis"
+              multiline
+              style={{ margin: 8 }}
+              name="synopsis"
+              helperText={errors.synopsis || `Caracteres restantes: ${(2000 - bookForm.synopsis.length)} `}
+              fullWidth
+              rows="4"
+              margin="normal"
+              InputLabelProps={{
+                shrink: true,
+              }}
+              inputProps={{
+                maxLength: 2000,
+              }}
+              onChange={({ target: { name, value } }) => setBookForm(name, value)}
+              required
+              error={errors.synopsis.length > 0}
+              value={bookForm.synopsis}
+            />
+            <LocalizationProvider dateAdapter={AdapterDateFns}>
+              <Grid container justifyContent="space-around">
+                <DatePicker
+                  label="Fecha de publicación"
+                  format="MM/dd/yyyy"
+                  value={bookForm.datePublished ? new Date(bookForm.datePublished) : null}
+                  onChange={(date) => setBookForm('datePublished', date)}
+                  slotProps={{
+                    textField: {
+                      margin: 'normal',
+                      id: 'date-picker-dialog',
+                      name: 'datePublished',
+                      required: true,
+                    },
+                  }}
+                />
+              </Grid>
+              <FormHelperText error>{errors.datePublished}</FormHelperText>
+            </LocalizationProvider>
+            {
+              registerBookResponse.message
+              && (
+                <Collapse in={open}>
+                  <Alert
+                    action={(
+                      <IconButton
+                        aria-label="close"
+                        color="inherit"
+                        size="small"
+                        onClick={() => {
+                          setOpen(false);
+                        }}
+                      >
+                        <CloseIcon fontSize="inherit" />
+                      </IconButton>
+                    )}
+                    variant="filled"
+                    severity={registerBookResponse.success ? 'success' : 'error'}
+                  >
+                    {registerBookResponse.message}
+                  </Alert>
+                </Collapse>
+              )
+            }
+          </StyledSecondColumnContainer>
+          <StyledThirdColumnContainer>
+            <GenresSelector
+              genreSelected={bookForm.genre}
+              onChange={(
+                { target: { name, value } }: React.ChangeEvent<HTMLInputElement>,
+              ) => setBookForm(name, value)}
+              errors={errors.genre}
+            />
+            <FormHelperText error>{errors.genre}</FormHelperText>
+            <TextField
+              id="standard-number"
+              label="Cantidad de páginas"
+              type="number"
+              name="pages"
+              size="small"
+              InputLabelProps={{
+                shrink: true,
+              }}
+              onChange={({ target: { name, value } }) => setBookForm(name, value)}
+              required
+              error={errors.pages.length > 0}
+              helperText={errors.pages}
+              value={bookForm.pages}
+            />
+            <FormatsCheckBoxSelector
+              formatsSelected={bookForm.formats}
+              errors={errors.formats}
+              options={['epub', 'papel', 'mobi', 'pdf', 'audiolibro']}
+              onChange={(
+                { target: { name, checked } }: React.ChangeEvent<HTMLInputElement>,
+              ) => setBookForm('formats',
+                checked
+                  ? [...bookForm.formats, name]
+                  : [...bookForm.formats.filter((format: string) => format !== name)])}
+              title="Formatos disponibles:"
+            />
+            <Button
+              disabled={loading}
+              variant="contained"
+              color="primary"
+              size="large"
+              onClick={submit}
+            >
+              {!book ? t('buttons.registerBook') : 'actualizar libro'}
+            </Button>
+
+          </StyledThirdColumnContainer>
+        </StyledFormContainer>
+      </Card>
+    </>
+  );
+};
+
+const StyledFormContainer = styledComponents(CardContent)`
+  @media (max-width: 375px) {
+    display: grid;
+    grid-template-columns: 1fr;
+  }
+  display: grid;
+  grid-template-columns: 0.70fr 2fr 1fr;
+  grid-gap: 2rem;
+`;
+
+const StyledFirstColumnContainer = styledComponents.div`
+  padding-top: 1rem;
+  display: grid;
+  grid-template-rows: 0.2fr 0.1fr 3fr;
+  justify-items: center;
+  grid-gap: 1rem;
+`;
+
+const StyledSecondColumnContainer = styledComponents.div`
+  display: grid;
+  grid-template-rows: 1fr;
+  grid-gap: 1rem;
+  padding: 3% 3% 3% 0;
+`;
+
+const StyledThirdColumnContainer = styledComponents.div`
+  padding-top: 1rem;
+  display: grid;
+  grid-template-rows: 1fr;
+  grid-gap: 1rem;
+`;
+
+const StyledImageContainer = styledComponents.img`
+  width: 70%;
+  border:2px solid #fff;
+  box-shadow: 10px 10px 5px #ccc;
+`;
+
+const StyledInputFile = styledComponents.input`
+  display: none;
+`;
+
+const StyledLabel = styledComponents.label`
+  color: ${(props) => props.theme.contrastText};
+  border-radius: 5px;
+  border:2px solid #fff;
+  box-shadow: 2px 2px 5px #ccc;
+  padding: 5%;
+  width: 70%;
+  align-self: flex-start;
+  text-align: center;
+  font-weight: 600;
+  background-color: ${(props) => props.theme.main};
+  &&:hover {
+    cursor: pointer;
+    background-color: ${(props) => props.theme.dark};
+    color: ${(props) => props.theme.main};
+  }
+`;
+
+export default AddBookSection;
