+++ src/views/AccountPage/sections/ProfileSection.tsx
@@ -1,0 +1,217 @@
+/* eslint-disable no-underscore-dangle */
+import React, { useContext, useState, useEffect } from 'react';
+import styledComponents from 'styled-components';
+import { useTranslation } from 'react-i18next';
+import {
+  Avatar,
+  Button,
+  Card,
+  CardContent,
+  Collapse,
+  IconButton,
+  TextField,
+} from '@mui/material';
+import CloseIcon from '@mui/icons-material/Close';
+import Alert from '@mui/material/Alert';
+import DeleteIcon from '@mui/icons-material/Delete';
+import UserContext from '../../../store/context/userContext/UserContext';
+import { UploadImagesInput, CountriesSelector, ModalDeleteAccount } from '../../../components';
+import { useForm, useUploadImages } from '../../../utils/customHooks';
+import { update as URL } from '../../../config/routes';
+import { Response } from '../../../interfaces/response';
+import SectionHeader from '../SectionHeader';
+
+const ProfileSection: React.FC = (): JSX.Element => {
+  const { t } = useTranslation();
+  const [response, setResponse] = useState<Response>({
+    success: undefined,
+    message: undefined,
+  });
+  const { user, setUserLogged } = useContext(UserContext);
+  const [updateForm, setUpdateForm, loadForm] = useForm(user);
+  const [avatarURL, uploadAvatar] = useUploadImages(updateForm.avatar);
+  const [open, setOpen] = useState(false);
+  const [openDeleteAccountModal, setDeleteAccountModal] = useState(false);
+  const [loading, setLoading] = useState(false);
+
+  useEffect(() => {
+    loadForm({ ...user });
+  }, [user._id]);
+
+  useEffect(() => {
+    setUpdateForm('avatar', avatarURL);
+  }, [avatarURL]);
+
+  const update = async (): Promise<void> => {
+    try {
+      setLoading(true);
+      const res = await fetch(URL, {
+        method: 'post',
+        mode: 'cors',
+        cache: 'no-cache',
+        credentials: 'include',
+        body: JSON.stringify({ ...updateForm }),
+        headers: {
+          'Content-Type': 'application/json',
+          'access-token': user.token,
+        },
+      });
+      const resJSON = await res.json();
+      const {
+        message,
+        success,
+        user: userUpdated,
+      } = resJSON;
+      setResponse({ message, success });
+      setUserLogged({ ...userUpdated, token: user.token });
+      setOpen(true);
+    } catch (error) {
+      setResponse(error);
+      setOpen(true);
+    } finally {
+      setLoading(false);
+    }
+  };
+
+  return (
+    <>
+      <SectionHeader title={t('titles.updateProfile')} />
+      <StyledCard>
+        <StyledAvatarContainer>
+          <UploadImagesInput
+            text="Sube tu Avatar"
+            onChange={(e: React.ChangeEvent<HTMLInputElement>) => uploadAvatar(e, 'avatars')}
+          />
+          {
+            updateForm.avatar
+            && <Avatar alt="avatar" src={updateForm.avatar} sx={{ width: 56, height: 56, alignContent: 'center', justifySelf: 'center', justifyContent: 'center' }} />
+          }
+        </StyledAvatarContainer>
+        <StyledContentContainer>
+          <StyledForm>
+            {
+              ['name', 'lastName', 'email'].map((text) => (
+                <TextField
+                  id="outlined-basic"
+                  label={t(`form.${text}`)}
+                  name={text}
+                  onChange={({ target: { name, value } }) => setUpdateForm(name, value)}
+                  variant={text === 'email' ? 'filled' : 'outlined'}
+                  defaultValue={user[text]}
+                  InputProps={{
+                    readOnly: (text === 'email'),
+                  }}
+                />
+              ))
+            }
+            <CountriesSelector
+              onChange={(
+                { target: { name, value } }: React.ChangeEvent<HTMLInputElement>,
+              ) => setUpdateForm(name, value)}
+              countrySelected={updateForm.country || user.country}
+            />
+          </StyledForm>
+          <StyledButton
+            disabled={loading}
+            variant="contained"
+            color="primary"
+            size="large"
+            onClick={update}
+          >
+            {loading ? t('buttons.saving') : t('buttons.update')}
+          </StyledButton>
+          <StyledResponseContainer>
+            {
+              response.message
+              && (
+                <Collapse in={open}>
+                  <Alert
+                    action={(
+                      <IconButton
+                        aria-label="close"
+                        color="inherit"
+                        size="small"
+                        onClick={() => {
+                          setOpen(false);
+                        }}
+                      >
+                        <CloseIcon fontSize="inherit" />
+                      </IconButton>
+                    )}
+                    variant="filled"
+                    severity={response.success ? 'success' : 'error'}
+                  >
+                    {t(`responses.${response.message}`)}
+                  </Alert>
+                </Collapse>
+              )
+            }
+          </StyledResponseContainer>
+        </StyledContentContainer>
+        <IconButton
+          aria-label="delete"
+          size="small"
+          color="secondary"
+          onClick={() => setDeleteAccountModal(true)}
+        >
+          <DeleteIcon fontSize="small" color="secondary" />
+          Eliminar mi cuenta
+        </IconButton>
+        <ModalDeleteAccount
+          open={openDeleteAccountModal}
+          onClose={() => setDeleteAccountModal(false)}
+        />
+      </StyledCard>
+    </>
+  );
+};
+
+const StyledCard = styledComponents(Card)`
+  display: grid;
+  grid-template-columns: 1fr 3fr;
+  @media (max-width: 375px) {
+    display: grid;
+    grid-template-columns: 1fr;
+  }
+`;
+
+const StyledAvatarContainer = styledComponents.div`
+  display: grid;
+  grid-template-rows: 0.5fr 2fr;
+  grid-gap: 0.5rem;
+  justify-content: center;
+  padding-top: 1rem;
+`;
+
+const StyledForm = styledComponents.form`
+  @media (max-width: 375px) {
+    display: grid;
+    grid-template-columns: 1fr;
+  }
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  grid-gap: 1rem;
+  width: 90%;
+  justify-self: center;
+`;
+
+const StyledButton = styledComponents(Button)`
+  @media (max-width: 375px) {
+    width: 50%;
+  }
+  width: 30%;
+  justify-self: center;
+`;
+
+const StyledContentContainer = styledComponents(CardContent)`
+  display: grid;
+  grid-template-rows: 1fr;
+  grid-gap: 1rem;
+`;
+
+const StyledResponseContainer = styledComponents.div`
+  width: 30%;
+  justify-self: center;
+`;
+
+export default ProfileSection;
